Added Suspense data fetching
