+++ src/index.css
@@ -118,3 +118,8 @@ body {
   box-shadow: 0 0 10px 10px hsla(150, 100%, 66%, .6);
   transform: scale(1.07);
 }
+
+.loading {
+  color: white;
+  font-size: 48px;
+}
